Test Setup for GaenV2Controller
